Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-animations/parsing/animation-timing-function-computed-expected.txt


PASS Property animation-timing-function value 'linear' computes to 'linear' 
PASS Property animation-timing-function value 'ease' computes to 'ease' 
PASS Property animation-timing-function value 'ease-in' computes to 'ease-in' 
PASS Property animation-timing-function value 'ease-out' computes to 'ease-out' 
PASS Property animation-timing-function value 'ease-in-out' computes to 'ease-in-out' 
PASS Property animation-timing-function value 'cubic-bezier(0.1, 0.2, 0.8, 0.9)' computes to 'cubic-bezier(0.1, 0.2, 0.8, 0.9)' 
PASS Property animation-timing-function value 'cubic-bezier(0, -2, 1, 3)' computes to 'cubic-bezier(0, -2, 1, 3)' 
PASS Property animation-timing-function value 'cubic-bezier(0, 0.7, 1, 1.3)' computes to 'cubic-bezier(0, 0.7, 1, 1.3)' 
PASS Property animation-timing-function value 'steps(4, start)' computes to 'steps(4, start)' 
FAIL Property animation-timing-function value 'steps(2, end)' computes to 'steps(2)' assert_equals: expected "steps(2)" but got "steps(2, end)"
FAIL Property animation-timing-function value 'steps(2, jump-start)' computes to 'steps(2, jump-start)' assert_true: 'steps(2, jump-start)' is a supported value for animation-timing-function. expected true got false
FAIL Property animation-timing-function value 'steps(2, jump-end)' computes to 'steps(2)' assert_true: 'steps(2, jump-end)' is a supported value for animation-timing-function. expected true got false
FAIL Property animation-timing-function value 'steps(2, jump-both)' computes to 'steps(2, jump-both)' assert_true: 'steps(2, jump-both)' is a supported value for animation-timing-function. expected true got false
FAIL Property animation-timing-function value 'steps(2, jump-none)' computes to 'steps(2, jump-none)' assert_true: 'steps(2, jump-none)' is a supported value for animation-timing-function. expected true got false
PASS Property animation-timing-function value 'linear, ease, linear' computes to 'linear, ease, linear' 
